refactor: move Gantt bar rendering from model to UI layer
